+++ novel_downloader/core/requesters/base_browser.py
@@ -111,7 +111,7 @@ class BaseBrowser(RequesterProtocol, abc.ABC):
         )
 
     @abc.abstractmethod
-    def get_book_info(self, book_id: str, wait_time: Optional[int] = None) -> str:
+    def get_book_info(self, book_id: str, wait_time: Optional[float] = None) -> str:
         """
         Fetch the raw HTML (or JSON) of the book info page.
 
@@ -123,7 +123,7 @@ class BaseBrowser(RequesterProtocol, abc.ABC):
 
     @abc.abstractmethod
     def get_book_chapter(
-        self, book_id: str, chapter_id: str, wait_time: Optional[int] = None
+        self, book_id: str, chapter_id: str, wait_time: Optional[float] = None
     ) -> str:
         """
         Fetch the raw HTML (or JSON) of a single chapter.
@@ -135,7 +135,7 @@ class BaseBrowser(RequesterProtocol, abc.ABC):
         """
         ...
 
-    def get_bookcase(self, wait_time: Optional[int] = None) -> str:
+    def get_bookcase(self, wait_time: Optional[float] = None) -> str:
         """
         Optional: Retrieve the HTML content of the authenticated user's bookcase page.
